refactor: move proof traits to `penumbra-proof-params`
