+++ crates/core/component/dex/src/swap_claim/action.rs
@@ -1,7 +1,8 @@
 use anyhow::Context;
 use penumbra_asset::Balance;
-use penumbra_crypto::{proofs::groth16::GROTH16_PROOF_LENGTH_BYTES, Nullifier};
+use penumbra_crypto::Nullifier;
 use penumbra_fee::Fee;
+use penumbra_proof_params::GROTH16_PROOF_LENGTH_BYTES;
 use penumbra_proto::{core::dex::v1alpha1 as pb, DomainType, TypeUrl};
 use penumbra_tct as tct;
