+++ crates/core/component/shielded-pool/src/spend/proof.rs
@@ -22,14 +22,12 @@ use penumbra_tct as tct;
 use rand_core::OsRng;
 
 use penumbra_asset::{balance, balance::commitment::BalanceCommitmentVar, Value};
-use penumbra_crypto::proofs::groth16::{
-    ParameterSetup, VerifyingKeyExt, GROTH16_PROOF_LENGTH_BYTES,
-};
 use penumbra_crypto::{note, Note, Nullifier, NullifierVar, Rseed};
 use penumbra_keys::keys::{
     AuthorizationKeyVar, IncomingViewingKeyVar, NullifierKey, NullifierKeyVar,
     RandomizedVerificationKey, SeedPhrase, SpendAuthRandomizerVar, SpendKey,
 };
+use penumbra_proof_params::{ParameterSetup, VerifyingKeyExt, GROTH16_PROOF_LENGTH_BYTES};
 
 /// Groth16 proof for spending existing notes.
 #[derive(Clone, Debug)]
@@ -328,11 +326,14 @@ impl TryFrom<pb::ZkSpendProof> for SpendProof {
 mod tests {
     use super::*;
     use ark_ff::UniformRand;
+    use ark_r1cs_std::prelude::Boolean;
     use decaf377::{Fq, Fr};
     use penumbra_asset::{asset, Value};
-    use penumbra_crypto::{Nullifier};
+    use penumbra_crypto::note::StateCommitment;
+    use penumbra_crypto::Nullifier;
     use penumbra_keys::{
         keys::{SeedPhrase, SpendKey},
+        Address,
     };
     use proptest::prelude::*;
 
@@ -340,7 +341,7 @@ mod tests {
     use penumbra_tct as tct;
     use rand_core::OsRng;
 
-    use penumbra_crypto::{ Note};
+    use penumbra_crypto::Note;
 
     use ark_ff::PrimeField;
 
@@ -767,6 +768,169 @@ mod tests {
             .expect("should be able to form proof");
 
             let proof_result = proof.verify(&vk, anchor, balance_commitment, nf, rk);
+            assert!(proof_result.is_ok());
+        }
+    }
+
+    struct MerkleProofCircuit {
+        /// Witness: Inclusion proof for the note commitment.
+        state_commitment_proof: tct::Proof,
+        /// Public input: The merkle root of the state commitment tree
+        pub anchor: tct::Root,
+    }
+
+    impl ConstraintSynthesizer<Fq> for MerkleProofCircuit {
+        fn generate_constraints(
+            self,
+            cs: ConstraintSystemRef<Fq>,
+        ) -> ark_relations::r1cs::Result<()> {
+            let merkle_path_var = tct::r1cs::MerkleAuthPathVar::new_witness(cs.clone(), || {
+                Ok(self.state_commitment_proof.clone())
+            })?;
+            let anchor_var = FqVar::new_input(cs.clone(), || Ok(Fq::from(self.anchor)))?;
+            let claimed_note_commitment =
+                note::StateCommitmentVar::new_witness(cs.clone(), || {
+                    Ok(self.state_commitment_proof.commitment())
+                })?;
+            let position_var = tct::r1cs::PositionVar::new_witness(cs.clone(), || {
+                Ok(self.state_commitment_proof.position())
+            })?;
+            let position_bits = position_var.inner.to_bits_le()?;
+            merkle_path_var.verify(
+                cs,
+                &Boolean::TRUE,
+                &position_bits,
+                anchor_var,
+                claimed_note_commitment.inner(),
+            )?;
+            Ok(())
+        }
+    }
+
+    impl ParameterSetup for MerkleProofCircuit {
+        fn generate_test_parameters() -> (ProvingKey<Bls12_377>, VerifyingKey<Bls12_377>) {
+            let seed_phrase = SeedPhrase::from_randomness([b'f'; 32]);
+            let sk_sender = SpendKey::from_seed_phrase(seed_phrase, 0);
+            let fvk_sender = sk_sender.full_viewing_key();
+            let ivk_sender = fvk_sender.incoming();
+            let (address, _dtk_d) = ivk_sender.payment_address(0u32.into());
+
+            let note = Note::from_parts(
+                address,
+                Value::from_str("1upenumbra").expect("valid value"),
+                Rseed([1u8; 32]),
+            )
+            .expect("can make a note");
+            let mut sct = tct::Tree::new();
+            let note_commitment = note.commit();
+            sct.insert(tct::Witness::Keep, note_commitment).unwrap();
+            let anchor = sct.root();
+            let state_commitment_proof = sct.witness(note_commitment).unwrap();
+            let circuit = MerkleProofCircuit {
+                state_commitment_proof,
+                anchor,
+            };
+            let (pk, vk) = Groth16::<Bls12_377, LibsnarkReduction>::circuit_specific_setup(
+                circuit, &mut OsRng,
+            )
+            .expect("can perform circuit specific setup");
+            (pk, vk)
+        }
+    }
+
+    fn make_random_note_commitment(address: Address) -> StateCommitment {
+        let note = Note::from_parts(
+            address,
+            Value::from_str("1upenumbra").expect("valid value"),
+            Rseed([1u8; 32]),
+        )
+        .expect("can make a note");
+        note.commit()
+    }
+
+    #[test]
+    fn merkle_proof_verification_succeeds() {
+        let (pk, vk) = MerkleProofCircuit::generate_prepared_test_parameters();
+        let mut rng = OsRng;
+
+        let seed_phrase = SeedPhrase::from_randomness([b'f'; 32]);
+        let sk_sender = SpendKey::from_seed_phrase(seed_phrase, 0);
+        let fvk_sender = sk_sender.full_viewing_key();
+        let ivk_sender = fvk_sender.incoming();
+        let (address, _dtk_d) = ivk_sender.payment_address(0u32.into());
+        // We will incrementally add notes to the state commitment tree, checking the merkle proofs verify
+        // at each step.
+        let mut sct = tct::Tree::new();
+
+        for _ in 0..5 {
+            let note_commitment = make_random_note_commitment(address);
+            sct.insert(tct::Witness::Keep, note_commitment).unwrap();
+            let anchor = sct.root();
+            let state_commitment_proof = sct.witness(note_commitment).unwrap();
+            let circuit = MerkleProofCircuit {
+                state_commitment_proof,
+                anchor,
+            };
+            let proof = Groth16::<Bls12_377, LibsnarkReduction>::prove(&pk, circuit, &mut rng)
+                .expect("should be able to form proof");
+
+            let proof_result = Groth16::<Bls12_377, LibsnarkReduction>::verify_with_processed_vk(
+                &vk,
+                &[Fq::from(anchor)],
+                &proof,
+            );
+            assert!(proof_result.is_ok());
+        }
+
+        sct.end_block().expect("can end block");
+        for _ in 0..100 {
+            let note_commitment = make_random_note_commitment(address);
+            sct.insert(tct::Witness::Forget, note_commitment).unwrap();
+        }
+
+        for _ in 0..5 {
+            let note_commitment = make_random_note_commitment(address);
+            sct.insert(tct::Witness::Keep, note_commitment).unwrap();
+            let anchor = sct.root();
+            let state_commitment_proof = sct.witness(note_commitment).unwrap();
+            let circuit = MerkleProofCircuit {
+                state_commitment_proof,
+                anchor,
+            };
+            let proof = Groth16::<Bls12_377, LibsnarkReduction>::prove(&pk, circuit, &mut rng)
+                .expect("should be able to form proof");
+
+            let proof_result = Groth16::<Bls12_377, LibsnarkReduction>::verify_with_processed_vk(
+                &vk,
+                &[Fq::from(anchor)],
+                &proof,
+            );
+            assert!(proof_result.is_ok());
+        }
+
+        sct.end_epoch().expect("can end epoch");
+        for _ in 0..100 {
+            let note_commitment = make_random_note_commitment(address);
+            sct.insert(tct::Witness::Forget, note_commitment).unwrap();
+        }
+
+        for _ in 0..5 {
+            let note_commitment = make_random_note_commitment(address);
+            sct.insert(tct::Witness::Keep, note_commitment).unwrap();
+            let anchor = sct.root();
+            let state_commitment_proof = sct.witness(note_commitment).unwrap();
+            let circuit = MerkleProofCircuit {
+                state_commitment_proof,
+                anchor,
+            };
+            let proof = Groth16::<Bls12_377, LibsnarkReduction>::prove(&pk, circuit, &mut rng)
+                .expect("should be able to form proof");
+
+            let proof_result = Groth16::<Bls12_377, LibsnarkReduction>::verify_with_processed_vk(
+                &vk,
+                &[Fq::from(anchor)],
+                &proof,
+            );
             assert!(proof_result.is_ok());
         }
     }
