--- crates/core/crypto/src/proofs.rs
@@ -1,1 +1,0 @@
-pub mod groth16;
