--- crates/core/crypto/src/proofs/groth16.rs
@@ -1,211 +1,0 @@
-mod traits;
-
-pub use traits::{ParameterSetup, ProvingKeyExt, VerifyingKeyExt};
-
-/// The length of our Groth16 proofs in bytes.
-pub const GROTH16_PROOF_LENGTH_BYTES: usize = 192;
-
-#[cfg(test)]
-mod tests {
-    use std::str::FromStr;
-
-    use super::*;
-    use crate::{Nullifier, Rseed};
-    use ark_ff::UniformRand;
-    use ark_groth16::{r1cs_to_qap::LibsnarkReduction, Groth16, ProvingKey, VerifyingKey};
-    use ark_r1cs_std::prelude::*;
-    use ark_relations::r1cs::{ConstraintSynthesizer, ConstraintSystemRef};
-    use ark_snark::SNARK;
-    use decaf377::{r1cs::FqVar, Bls12_377, Fq, Fr};
-    use penumbra_asset::{asset, Balance, Value};
-    use penumbra_keys::{
-        keys::{SeedPhrase, SpendKey},
-        Address,
-    };
-    use proptest::prelude::*;
-
-    use decaf377_rdsa::{SpendAuth, VerificationKey};
-    use penumbra_proto::core::crypto::v1alpha1 as pb;
-    use penumbra_tct as tct;
-    use rand_core::OsRng;
-    use tct::StateCommitment;
-
-    use crate::{note, Note};
-
-    use ark_ff::PrimeField;
-
-    fn fq_strategy() -> BoxedStrategy<Fq> {
-        any::<[u8; 32]>()
-            .prop_map(|bytes| Fq::from_le_bytes_mod_order(&bytes[..]))
-            .boxed()
-    }
-
-    fn fr_strategy() -> BoxedStrategy<Fr> {
-        any::<[u8; 32]>()
-            .prop_map(|bytes| Fr::from_le_bytes_mod_order(&bytes[..]))
-            .boxed()
-    }
-
-    struct MerkleProofCircuit {
-        /// Witness: Inclusion proof for the note commitment.
-        state_commitment_proof: tct::Proof,
-        /// Public input: The merkle root of the state commitment tree
-        pub anchor: tct::Root,
-    }
-
-    impl ConstraintSynthesizer<Fq> for MerkleProofCircuit {
-        fn generate_constraints(
-            self,
-            cs: ConstraintSystemRef<Fq>,
-        ) -> ark_relations::r1cs::Result<()> {
-            let merkle_path_var = tct::r1cs::MerkleAuthPathVar::new_witness(cs.clone(), || {
-                Ok(self.state_commitment_proof.clone())
-            })?;
-            let anchor_var = FqVar::new_input(cs.clone(), || Ok(Fq::from(self.anchor)))?;
-            let claimed_note_commitment =
-                note::StateCommitmentVar::new_witness(cs.clone(), || {
-                    Ok(self.state_commitment_proof.commitment())
-                })?;
-            let position_var = tct::r1cs::PositionVar::new_witness(cs.clone(), || {
-                Ok(self.state_commitment_proof.position())
-            })?;
-            let position_bits = position_var.inner.to_bits_le()?;
-            merkle_path_var.verify(
-                cs,
-                &Boolean::TRUE,
-                &position_bits,
-                anchor_var,
-                claimed_note_commitment.inner(),
-            )?;
-            Ok(())
-        }
-    }
-
-    impl ParameterSetup for MerkleProofCircuit {
-        fn generate_test_parameters() -> (ProvingKey<Bls12_377>, VerifyingKey<Bls12_377>) {
-            let seed_phrase = SeedPhrase::from_randomness([b'f'; 32]);
-            let sk_sender = SpendKey::from_seed_phrase(seed_phrase, 0);
-            let fvk_sender = sk_sender.full_viewing_key();
-            let ivk_sender = fvk_sender.incoming();
-            let (address, _dtk_d) = ivk_sender.payment_address(0u32.into());
-
-            let note = Note::from_parts(
-                address,
-                Value::from_str("1upenumbra").expect("valid value"),
-                Rseed([1u8; 32]),
-            )
-            .expect("can make a note");
-            let mut sct = tct::Tree::new();
-            let note_commitment = note.commit();
-            sct.insert(tct::Witness::Keep, note_commitment).unwrap();
-            let anchor = sct.root();
-            let state_commitment_proof = sct.witness(note_commitment).unwrap();
-            let circuit = MerkleProofCircuit {
-                state_commitment_proof,
-                anchor,
-            };
-            let (pk, vk) = Groth16::<Bls12_377, LibsnarkReduction>::circuit_specific_setup(
-                circuit, &mut OsRng,
-            )
-            .expect("can perform circuit specific setup");
-            (pk, vk)
-        }
-    }
-
-    fn make_random_note_commitment(address: Address) -> StateCommitment {
-        let note = Note::from_parts(
-            address,
-            Value::from_str("1upenumbra").expect("valid value"),
-            Rseed([1u8; 32]),
-        )
-        .expect("can make a note");
-        note.commit()
-    }
-
-    #[test]
-    fn merkle_proof_verification_succeeds() {
-        let (pk, vk) = MerkleProofCircuit::generate_prepared_test_parameters();
-        let mut rng = OsRng;
-
-        let seed_phrase = SeedPhrase::from_randomness([b'f'; 32]);
-        let sk_sender = SpendKey::from_seed_phrase(seed_phrase, 0);
-        let fvk_sender = sk_sender.full_viewing_key();
-        let ivk_sender = fvk_sender.incoming();
-        let (address, _dtk_d) = ivk_sender.payment_address(0u32.into());
-        // We will incrementally add notes to the state commitment tree, checking the merkle proofs verify
-        // at each step.
-        let mut sct = tct::Tree::new();
-
-        for _ in 0..5 {
-            let note_commitment = make_random_note_commitment(address);
-            sct.insert(tct::Witness::Keep, note_commitment).unwrap();
-            let anchor = sct.root();
-            let state_commitment_proof = sct.witness(note_commitment).unwrap();
-            let circuit = MerkleProofCircuit {
-                state_commitment_proof,
-                anchor,
-            };
-            let proof = Groth16::<Bls12_377, LibsnarkReduction>::prove(&pk, circuit, &mut rng)
-                .expect("should be able to form proof");
-
-            let proof_result = Groth16::<Bls12_377, LibsnarkReduction>::verify_with_processed_vk(
-                &vk,
-                &[Fq::from(anchor)],
-                &proof,
-            );
-            assert!(proof_result.is_ok());
-        }
-
-        sct.end_block().expect("can end block");
-        for _ in 0..100 {
-            let note_commitment = make_random_note_commitment(address);
-            sct.insert(tct::Witness::Forget, note_commitment).unwrap();
-        }
-
-        for _ in 0..5 {
-            let note_commitment = make_random_note_commitment(address);
-            sct.insert(tct::Witness::Keep, note_commitment).unwrap();
-            let anchor = sct.root();
-            let state_commitment_proof = sct.witness(note_commitment).unwrap();
-            let circuit = MerkleProofCircuit {
-                state_commitment_proof,
-                anchor,
-            };
-            let proof = Groth16::<Bls12_377, LibsnarkReduction>::prove(&pk, circuit, &mut rng)
-                .expect("should be able to form proof");
-
-            let proof_result = Groth16::<Bls12_377, LibsnarkReduction>::verify_with_processed_vk(
-                &vk,
-                &[Fq::from(anchor)],
-                &proof,
-            );
-            assert!(proof_result.is_ok());
-        }
-
-        sct.end_epoch().expect("can end epoch");
-        for _ in 0..100 {
-            let note_commitment = make_random_note_commitment(address);
-            sct.insert(tct::Witness::Forget, note_commitment).unwrap();
-        }
-
-        for _ in 0..5 {
-            let note_commitment = make_random_note_commitment(address);
-            sct.insert(tct::Witness::Keep, note_commitment).unwrap();
-            let anchor = sct.root();
-            let state_commitment_proof = sct.witness(note_commitment).unwrap();
-            let circuit = MerkleProofCircuit {
-                state_commitment_proof,
-                anchor,
-            };
-            let proof = Groth16::<Bls12_377, LibsnarkReduction>::prove(&pk, circuit, &mut rng)
-                .expect("should be able to form proof");
-
-            let proof_result = Groth16::<Bls12_377, LibsnarkReduction>::verify_with_processed_vk(
-                &vk,
-                &[Fq::from(anchor)],
-                &proof,
-            );
-            assert!(proof_result.is_ok());
-        }
-    }
-}
