+++ crates/crypto/proof-params/src/traits.rs
